pyproject.toml
[project]
name = "avtomail"
version = "0.1.0"
description = "Automated multilingual email assistant backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "asyncpg>=0.28",
    "alembic>=1.13",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "langdetect>=1.0.9",
    "httpx>=0.27",
    "email-validator>=2.1",
    "passlib[bcrypt]>=1.7",
    "python-jose[cryptography]>=3.3",
    "apscheduler>=3.10",
    "tenacity>=8.2",
    "python-multipart>=0.0.9",
    "typer>=0.9",
    "jinja2>=3.1",
    "aiofiles>=23.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "aiosqlite>=0.19",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.8"
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
[tool.setuptools.packages.find]
where = ["backend"]


